debut d'integration en angular (page authentification et rajout des service web size)

+++ src/main/java/com/gestion/incident/service/ClientService.java
@@ -1,12 +1,14 @@
 package com.gestion.incident.service;
 
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.gestion.incident.bean.Client;
+import com.gestion.incident.exception.IncidentException;
 import com.gestion.incident.repositories.ClientRepository;
 
 
@@ -15,6 +17,7 @@ public class ClientService {
 
 	@Autowired
 	public ClientRepository clientRepository;
+	
 
 //déclaration des attributs	
 	//Calendar calendar = new GregorianCalendar(2014, 9, 31);
@@ -38,15 +41,25 @@ public class ClientService {
 		return "le client a été ajouter avec succès";
 	}
 	
-	public void UpdateClient(Client client){
+	public void UpdateClient(Client client) throws IncidentException{
 	if(clientRepository.exists(client.getId())){
 		clientRepository.save(client);
-	}
+	}else 
+		throw new IncidentException("le client n'existe pas pour une mise à jour");
 }
 	
-	public void DeleteClient(String idclient) {
+
+	public void DeleteClient(String idclient) throws IncidentException {
+		if(clientRepository.exists("idclient")){
 		clientRepository.delete(idclient);
+		}else
+			throw new IncidentException("l'id du client à supprimer n'existe pas");
 		//clients.removeIf(client -> client.getId().equals(idclient));
+	}
+	
+	public int NumberClients(){
+		ArrayList<Client> ListClient = (ArrayList<Client>)clientRepository.findAll();
+		return ListClient.size();
 	}
 	
 }
